+++ app/src/main/java/com/crost/aurorabzalarm/data/model/AceEpamData.kt
@@ -9,8 +9,8 @@ import com.crost.aurorabzalarm.utils.Constants.EPAM_TABLE_NAME
 data class AceEpamData(
     @PrimaryKey(autoGenerate = false)
     var datetime: Long, //source format: YYYY-MM-DD_HH:MM
-    var density: Float,
-    var speed: Float,
-    var temp: Float,
+    var density: Double,
+    var speed: Double,
+    var temp: Double,
 
 )
