+++ app/src/main/java/com/crost/aurorabzalarm/network/parser/converter/EpamDataConverter.kt
@@ -15,15 +15,15 @@ class EpamDataConverter {
     {
         val convertedDataTable = mutableListOf<MutableMap<String, Any>>()
 
-        var protonDensity: Float
-        var bulkSpeed: Float
-        var ionTemperature: Float
+        var protonDensity: Double
+        var bulkSpeed: Double
+        var ionTemperature: Double
 
         for (dataMap in dataTable){
             val row = mutableMapOf<String, Any>()
-            protonDensity = -9999.9f
-            bulkSpeed = -9999.9f
-            ionTemperature = -1.00e+05f
+            protonDensity = -9999.9
+            bulkSpeed = -9999.9
+            ionTemperature = -1.00e+05
 
             val datetime = try {
                 convertToLocalEpochMillis(
@@ -39,9 +39,9 @@ class EpamDataConverter {
             }
 
             try {
-                protonDensity = dataMap["ProtonDensity"]?.toFloat()!!
-                bulkSpeed = dataMap["BulkSpeed"]?.toFloat()!!
-                ionTemperature = dataMap["IonTemperature"]?.toFloat()!!
+                protonDensity = dataMap["ProtonDensity"]?.toDouble()!!
+                bulkSpeed = dataMap["BulkSpeed"]?.toDouble()!!
+                ionTemperature = dataMap["IonTemperature"]?.toDouble()!!
 //                Log.d("EpamDataConverter",
 //                    "values: $protonDensity, $bulkSpeed, $ionTemperature")
             } catch (e: Exception){
